requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.0.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
boto3==1.26.76
botocore==1.29.76
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cinemagoer==2023.5.1
click==8.1.6
comm==0.1.3
dataclasses-json==0.5.13
debugpy==1.6.7
decorator==5.1.1
deeplake==3.6.14
dill==0.3.7
entrypoints==0.4
executing==1.2.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
humbug==0.3.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.247
langsmith==0.0.15
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.7
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathos==0.3.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.9.1
pox==0.3.3
ppft==1.7.6.7
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.0
requests==2.31.0
rich==13.5.0
rpds-py==0.9.2
s3transfer==0.6.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.19
stack-data==0.6.2
streamlit==1.25.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
wcwidth==0.2.6
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
